Try to authenticate with app via JSON Web Token (JWT).
This failed miserably with a bad credentials error. I tried both
with conda-forge PyGithub (1.43.6) and the latest release 1.51 with
integrations support:

pip install PyGithub['integrations']

Frustratingly there doesn't seem to be a way to specify the extra
integrations support in environment.yml. And it's not included in
`pip freeze` or `conda env export`.

Using PyGithub as a GitHub App is essentially undocumented. I've had to
piece together clues from the Issues, Pull Requests, and tests. It says
that that part of the API is supported:

https://pygithub.readthedocs.io/en/latest/github_objects/Installation.html

But I can't figure out how to authenticate. Passing the jwt token from
my app to the main class didn't work unfortunately.

At some point I'm just going to have to use pyjwt and requests directly.
At least that way I can follow the GitHub API documentation instead of
playing detective.